+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-# Microsoft 9.0
+# Microsoft 10.0
 
-- 将`Rougamo.Extensions.DependencyInjection.Microsoft`更新到`9.0.0`。
+- 将`Rougamo.Extensions.DependencyInjection.Microsoft`更新到`10.0.0`
+- 为`MethodContext`新增`GetKeyedService`系列方法